refactor(football-data): use version 4 of api

+++ src/domain/mod.rs
@@ -9,7 +9,7 @@ pub struct Competition {
 
 #[derive(Debug)]
 pub struct FixtureCollection {
-    pub count: u32,
+    pub count: u8,
     pub matches: Vec<Match>,
 }
